+++ src/com/automates/automate/sqlite/SQLiteDBManager.java
@@ -141,7 +141,7 @@ public class SQLiteDBManager extends SQLiteOpenHelper implements PatternManager{
         pattern.setActionCategory(cursor.getString(1));
         pattern.setAction(cursor.getString(2));
         pattern.setTime(Integer.parseInt(cursor.getString(3)));
-        pattern.setActualTime(Integer.parseInt(cursor.getString(4)));
+        pattern.setActualTime(Long.parseLong(cursor.getString(4)));
         pattern.setDay(Integer.parseInt(cursor.getString(5)));
         pattern.setLocation(cursor.getString(6));
         pattern.setWifi(cursor.getString(7));
@@ -175,7 +175,7 @@ public class SQLiteDBManager extends SQLiteOpenHelper implements PatternManager{
                 pattern.setActionCategory(cursor.getString(1));
                 pattern.setAction(cursor.getString(2));
                 pattern.setTime(Integer.parseInt(cursor.getString(3)));
-                pattern.setActualTime(Integer.parseInt(cursor.getString(4)));
+                pattern.setActualTime(Long.parseLong(cursor.getString(4)));
                 pattern.setDay(Integer.parseInt(cursor.getString(5)));
                 pattern.setLocation(cursor.getString(6));
                 pattern.setWifi(cursor.getString(7));
